hapus audit + perbaikan dashboard

+++ app/Http/Controllers/pemilik.php
@@ -8,6 +8,9 @@ use Illuminate\Http\Request;
 use App\Models\logadmin;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Redirect;
+
+use function PHPUnit\Framework\isNull;
 
 class pemilik extends Controller
 {
@@ -47,13 +50,16 @@ class pemilik extends Controller
             ->where('level', '>=', 2);
         //YANG DIUBAH
         // return view("pemilik.karyawan", ['data' => $data ]);
-        return view("pemilik.karyawan", ['data' => $data, 'data1' => pabrik::all() ]);
+        return view("pemilik.karyawan", ['data' => $data, 'data1' => pabrik::all()]);
         //END
     }
 
     public function hapus_karyawan(Request $req)
     {
         // dd($req->id);
+        if (isNull($req['_token'])) {
+            return Redirect::back();
+        }
         $post = user::all()->where('id',  $req->id)->each->delete();
         $pabrik = Auth::user()->pabrik;
         $data = user::all()->where('pabrik', $pabrik)
@@ -61,14 +67,14 @@ class pemilik extends Controller
 
         if ($req->level == 2) {
             $in = 'penanggung jawab teknis';
-        } elseif($req->level == 3) {
+        } elseif ($req->level == 3) {
             $in = 'Pelaksana';
         }
-        
+
         if ($post) {
             $log = [
-                'log_isi' => session()->get('pabrik').' <b> Menghapus '. $req->namadepan . ' ' .  $req->namabelakang . ' </b> &nbsp dari '. $in,
-                'log_pabrik' => session()->get('pabrik'), 
+                'log_isi' => session()->get('pabrik') . ' <b> Menghapus ' . $req->namadepan . ' ' .  $req->namabelakang . ' </b> &nbsp dari ' . $in,
+                'log_pabrik' => session()->get('pabrik'),
                 'log_waktu' => date('Y-m-d H:i:s'),
                 'id_pabrik' => Auth::user()->pabrik
             ];
@@ -78,8 +84,8 @@ class pemilik extends Controller
         } else {
             return redirect('/karyawan')->with('error', 'Gagal dihapus!');
         }
-        
-        
+
+
         // return view("pemilik.karyawan", ['data' => $data]);
     }
 
@@ -92,21 +98,21 @@ class pemilik extends Controller
         $pabrik = Auth::user()->pabrik;
         $data = user::all()->where('pabrik', $pabrik)
             ->where('level', '>=', 2);
-            // dd($data);
+        // dd($data);
 
-            if ($req->posisi == 2) {
-                $in = 'penanggung jawab teknis';
-            } elseif($req->posisi == 3) {
-                $in = 'Pelaksana';
-            }
-            
-            $log = [
-                'log_isi' => session()->get('pabrik').' <b> Mengubah jabatan '. $req->nama . ' </b> &nbsp menjadi '. $in,
-                'log_pabrik' => session()->get('pabrik'), 
-                'log_waktu' => date('Y-m-d H:i:s'),
-                'id_pabrik' => Auth::user()->pabrik
-            ];
-            logadmin::insert($log);
+        if ($req->posisi == 2) {
+            $in = 'penanggung jawab teknis';
+        } elseif ($req->posisi == 3) {
+            $in = 'Pelaksana';
+        }
+
+        $log = [
+            'log_isi' => session()->get('pabrik') . ' <b> Mengubah jabatan ' . $req->nama . ' </b> &nbsp menjadi ' . $in,
+            'log_pabrik' => session()->get('pabrik'),
+            'log_waktu' => date('Y-m-d H:i:s'),
+            'id_pabrik' => Auth::user()->pabrik
+        ];
+        logadmin::insert($log);
 
         return view("pemilik.karyawan", ['data' => $data]);
     }
@@ -137,7 +143,7 @@ class pemilik extends Controller
     {
         // dd($req);
         $user = audit::all()->where("audit_pabrik", $req->pabrik)
-            
+
             ->where("audit_laporan", $req->laporan)
             ->where("audit_id", $req->no)->first()->update([
                 'audit_status' => 1
@@ -156,13 +162,11 @@ class pemilik extends Controller
         return redirect('bos_audit')->with('success', 'Data Berhasil Dihapus!');
     }
 
-    public function log () {
+    public function log()
+    {
         // dd('halo');
         // Posts::orderBy('created_at', 'desc')->get();
         // return view('layout.log', ['dataLog' => DB::select('select * from logs')]);
         return view('layout.logadmin', ['dataLog' => Logadmin::all()->where('id_pabrik', Auth::user()->pabrik)]);
-        
     }
-
-
 }
